# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_NR3101_go-ecom-project_internal_dto.UpdateProfileRequest:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
    required:
    - first_name
    - last_name
    type: object
  github_com_NR3101_go-ecom-project_internal_dto.UserResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_name:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
...
  github_com_NR3101_go-ecom-project_internal_utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: no-email@enail.com
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for an e-commerce application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swaager.io/terms/
  title: E-commerce API
  version: "1.0"
paths:
...
  /users/profile:
    get:
      consumes:
      - application/json
      description: Get the profile information of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User profile retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.UserResponse'
              type: object
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update the profile information of the authenticated user
      parameters:
      - description: User profile update data
        in: body
        name: updateProfileRequest
        required: true
        schema:
          $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User profile updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.UserResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - User
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {token}" to authenticate requests
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"